Updated Tutorial Section
The Tutorial section as of now is a full walkthrough section. But we
will have two tutorials in the future. One will be a full walkthrough
(in detail explanations of what to do). The other will be a quick start
guide so that users can download the repo and get started right away.